# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Transit-AI Backend",
    "description": "Прогноз пассажиропотока трамваев Москвы (hackathon Transit-AI).",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/v1/predictions/db/{route_id}": {
      "get": {
        "tags": [
          "predictions-db"
        ],
        "summary": "Predictions for route (from DB with feature_set / zeros filters)",
        "description": "Возвращает прогнозы из БД с фильтрацией.\n\nЕсли filters=None — использует дефолты из feature_toggles + zero_overrides.",
        "operationId": "get_predictions_db_api_v1_predictions_db__route_id__get",
        "parameters": [
          {
            "name": "route_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Route Id"
            }
          },
          {
            "name": "from",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date-time",
              "title": "To"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "feature_set",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature Set"
            }
          },
          {
            "name": "zeros_applied",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zeros Applied"
            }
          },
          {
            "name": "prefer_active",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "description": "T-230: игнорировать coef/feature_set/zeros и взять параметры активного набора (фолбэк фронта, когда по слайдерам нет данных).",
              "default": false,
              "title": "Prefer Active"
            },
            "description": "T-230: игнорировать coef/feature_set/zeros и взять параметры активного набора (фолбэк фронта, когда по слайдерам нет данных)."
          },
          {
            "name": "horizon",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "day",
                "month",
                "year"
              ],
              "type": "string",
              "default": "day",
              "title": "Horizon"
            }
          },
          {
            "name": "granularity",
            "in": "query",
            "required": false,
            "schema": {
              "enum": [
                "hour",
                "day",
                "month"
              ],
              "type": "string",
              "default": "hour",
              "title": "Granularity"
            }
          },
          {
            "name": "coef_weather",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 3,
              "minimum": 0,
              "default": 1.0,
              "title": "Coef Weather"
            }
          },
          {
            "name": "coef_event",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 3,
              "minimum": 0,
              "default": 1.0,
              "title": "Coef Event"
            }
          },
          {
            "name": "coef_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 3,
              "minimum": 0,
              "default": 1.0,
              "title": "Coef Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PredictionsDBResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/export.csv": {
      "get": {
        "tags": [
          "predictions-db"
        ],
        "summary": "Export predictions as CSV (default params = best submission)",
        "description": "Возвращает CSV (route;date;hour;prediction).\n\nDefault params: zeros_applied=True, feature_set=with_all — соответствует\nbest submission (F-083: 0.83455 platform score).\n\nФормат совместим с submission pipeline (clinerule 23):\n  - separator `;`\n  - колонки: route, date (YYYY-MM-DD), hour (0-23), prediction\n  - header НЕ включён (для совместимости с ml platform scoring)\n\nВалидация (clinerule 23):\n  - coef_weather/event/season ∈ [0, 3] (FastAPI Query ge/le → 422)\n  - from_date < to_date (HTTPException 400)",
        "operationId": "export_predictions_csv_api_v1_predictions_export_csv_get",
        "parameters": [
          {
            "name": "from_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": "2025-11-01T00:00:00Z",
              "title": "From Date"
            }
          },
          {
            "name": "to_date",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": "2025-12-31T23:00:00Z",
              "title": "To Date"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "feature_set",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature Set"
            }
          },
          {
            "name": "zeros_applied",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zeros Applied"
            }
          },
          {
            "name": "coef_weather",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 3,
              "minimum": 0,
              "default": 1.0,
              "title": "Coef Weather"
            }
          },
          {
            "name": "coef_event",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 3,
              "minimum": 0,
              "default": 1.0,
              "title": "Coef Event"
            }
          },
          {
            "name": "coef_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "maximum": 3,
              "minimum": 0,
              "default": 1.0,
              "title": "Coef Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/predictions/export.xlsx": {
      "get": {
        "tags": [
          "predictions-db"
        ],
        "summary": "Export predictions as XLSX (T-206: альтернатива CSV для аналитиков)",
        "description": "Возвращает XLSX (route, date, hour, prediction + coef колонки).\n\nУдобно для аналитиков, которые работают в Excel/LibreOffice.\nDefault params = best submission (F-083, 0.83455).",
        "operationId": "export_predictions_xlsx_api_v1_predictions_export_xlsx_get",
        "parameters": [
          {
            "name": "from",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": "2025-11-01T00:00:00Z",
              "title": "From"
            }
          },
          {
            "name": "to",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "date-time",
              "default": "2025-12-31T23:00:00Z",
              "title": "To"
            }
          },
          {
            "name": "model_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Model Id"
            }
          },
          {
            "name": "feature_set",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Feature Set"
            }
          },
          {
            "name": "zeros_applied",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "boolean"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Zeros Applied"
            }
          },
          {
            "name": "coef_weather",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Coef Weather"
            }
          },
          {
            "name": "coef_event",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Coef Event"
            }
          },
          {
            "name": "coef_season",
            "in": "query",
            "required": false,
            "schema": {
              "type": "number",
              "default": 1.0,
              "title": "Coef Season"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PredictionPointDB": {
        "properties": {
          "period_start": {
            "type": "string",
            "format": "date-time",
            "title": "Period Start"
          },
          "period_end": {
            "type": "string",
            "format": "date-time",
            "title": "Period End"
          },
          "value": {
            "type": "number",
            "title": "Value"
          },
          "lower": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Lower"
          },
          "upper": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Upper"
          },
          "horizon": {
            "type": "string",
            "title": "Horizon"
          },
          "granularity": {
            "type": "string",
            "title": "Granularity"
          },
          "feature_set": {
            "type": "string",
            "title": "Feature Set"
          },
          "zeros_applied": {
            "type": "boolean",
            "title": "Zeros Applied"
          },
          "coef_weather": {
            "type": "number",
            "title": "Coef Weather"
          },
          "coef_event": {
            "type": "number",
            "title": "Coef Event"
          },
          "coef_season": {
            "type": "number",
            "title": "Coef Season"
          }
        },
        "type": "object",
        "required": [
          "period_start",
          "period_end",
          "value",
          "horizon",
          "granularity",
          "feature_set",
          "zeros_applied",
          "coef_weather",
          "coef_event",
          "coef_season"
        ],
        "title": "PredictionPointDB",
        "description": "Одна точка прогноза из БД."
      },
...
      "PredictionsDBResponse": {
        "properties": {
          "route_id": {
            "type": "integer",
            "title": "Route Id"
          },
          "from_date": {
            "type": "string",
            "format": "date-time",
            "title": "From Date"
          },
          "to_date": {
            "type": "string",
            "format": "date-time",
            "title": "To Date"
          },
          "model_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Model Id"
          },
          "feature_set": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Feature Set"
          },
          "zeros_applied": {
            "anyOf": [
              {
                "type": "boolean"
              },
              {
                "type": "null"
              }
            ],
            "title": "Zeros Applied"
          },
          "points": {
            "items": {
              "$ref": "#/components/schemas/PredictionPointDB"
            },
            "type": "array",
            "title": "Points"
          }
        },
        "type": "object",
        "required": [
          "route_id",
          "from_date",
          "to_date",
          "points"
        ],
        "title": "PredictionsDBResponse",
        "description": "Ответ /predictions/db/{route_id}."
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    }
  }
}